Grouped execution setup and resetting logic
